+++ src/test/java/com/linbit/extproc/utils/TestExtCmd.java
@@ -1,31 +1,30 @@
 package com.linbit.extproc.utils;
 
-import static org.junit.Assert.fail;
-
 import com.linbit.ChildProcessTimeoutException;
 import com.linbit.extproc.ExtCmd;
-import com.linbit.linstor.logging.StdErrorReporter;
+import com.linbit.linstor.logging.ErrorReporter;
 import com.linbit.timer.Action;
 import com.linbit.timer.GenericTimer;
 
 import java.io.IOException;
 import java.lang.ProcessBuilder.Redirect;
-import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Map;
 import java.util.Map.Entry;
 
+import static org.junit.Assert.fail;
+
 public class TestExtCmd extends ExtCmd
 {
     private BehaviorManager behaviorMgr = new BehaviorManager();
 
-    public TestExtCmd()
+    public TestExtCmd(ErrorReporter errorReporterRef)
     {
         super(
-            new GenericTimer<String, Action<String>>(),
-            new StdErrorReporter("LINSTOR-UNITTESTS", Paths.get("build/test-logs"), true, "", null, null, () -> null)
+            new GenericTimer<>(),
+            errorReporterRef
         );
     }