+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed race condition during "linstor node reconnect" leading to reconnect-cycle
 - Fixed duplicate connections to same satellite (older connection gets dropped now)
+- Fixed problems with JUnit ProcCryptoUtilsTest
 
 ## [1.31.0] - 2025-04-08